+++ src/main/java/de/uib/configed/tree/AbstractGroupTree.java
@@ -134,7 +134,7 @@ public abstract class AbstractGroupTree extends JTree implements TreeSelectionLi
 		groups.clear();
 		rootNode.removeAllChildren();
 		groupNodeFullList.removeAllChildren();
-		model.nodeStructureChanged(groupNodeGroups);
+		model.nodeStructureChanged(rootNode);
 		createTree();
 
 		removeTreeSelectionListener(this);